+++ node/silkworm/db/access_layer_test.cpp
@@ -27,7 +27,7 @@
 #include <silkworm/db/prune_mode.hpp>
 #include <silkworm/db/tables.hpp>
 #include <silkworm/execution/execution.hpp>
-#include <silkworm/stagedsync/common.hpp>
+#include <silkworm/stagedsync/stage.hpp>
 #include <silkworm/stagedsync/stage_history_index.hpp>
 
 namespace silkworm {
