+++ node/silkworm/stagedsync/stage_bodies.hpp
@@ -21,7 +21,7 @@
 #include <silkworm/downloader/block_exchange.hpp>
 #include <silkworm/downloader/internals/types.hpp>
 #include <silkworm/downloader/messages/internal_message.hpp>
-#include <silkworm/stagedsync/common.hpp>
+#include <silkworm/stagedsync/stage.hpp>
 
 namespace silkworm::stagedsync {
