+++ node/silkworm/stagedsync/stage_execution.hpp
@@ -21,7 +21,7 @@
 #include <silkworm/consensus/engine.hpp>
 #include <silkworm/execution/analysis_cache.hpp>
 #include <silkworm/execution/evm.hpp>
-#include <silkworm/stagedsync/common.hpp>
+#include <silkworm/stagedsync/stage.hpp>
 
 namespace silkworm::stagedsync {
