+++ node/silkworm/stagedsync/stage_finish.hpp
@@ -16,7 +16,7 @@
 
 #pragma once
 
-#include <silkworm/stagedsync/common.hpp>
+#include <silkworm/stagedsync/stage.hpp>
 
 namespace silkworm::stagedsync {
