+++ node/silkworm/stagedsync/stage_history_index.hpp
@@ -17,7 +17,7 @@
 #pragma once
 
 #include <silkworm/db/bitmap.hpp>
-#include <silkworm/stagedsync/common.hpp>
+#include <silkworm/stagedsync/stage.hpp>
 
 namespace silkworm::stagedsync {
