+++ node/silkworm/stagedsync/stage_senders.hpp
@@ -16,7 +16,7 @@
 
 #pragma once
 
-#include <silkworm/stagedsync/common.hpp>
+#include <silkworm/stagedsync/stage.hpp>
 #include <silkworm/stagedsync/stage_senders/recovery_farm.hpp>
 
 namespace silkworm::stagedsync {
