+++ node/silkworm/stagedsync/sync_loop.hpp
@@ -22,7 +22,7 @@
 #include <silkworm/common/stopwatch.hpp>
 #include <silkworm/concurrency/worker.hpp>
 #include <silkworm/downloader/block_exchange.hpp>
-#include <silkworm/stagedsync/common.hpp>
+#include <silkworm/stagedsync/stage.hpp>
 
 namespace silkworm::stagedsync {
